회원가입할때 인벤토리 부분만 고치면 될듯

+++ ipmall/src/main/java/kr/co/ipmall/service/inventoryService/InventoryServiceImpl.java
@@ -14,7 +14,7 @@ public class InventoryServiceImpl implements InventoryService {
 	@Resource(name="inventoryDAO")
 	private InventoryDAO inventoryDAO;
 	
-	public void createInventory(int customer_no) {
+	public void createInventory(Long customer_no) {
 		inventoryDAO.insert("inventory.insertEmptyInventory", customer_no);
 	}
